+++ commands/utils/packageUpdaters/npmpackagehandler.go
@@ -1,4 +1,4 @@
-package utils
+package packageUpdaters
 
 type NpmPackageHandler struct {
 	GenericPackageHandler
